fix：将long改为int

+++ src/main/java/com/dd/ai_smart_course/controller/AnalysisController.java
@@ -23,8 +23,8 @@ public class AnalysisController {
     private JwtTokenUtil jwtTokenUtil;
 
     // 暂时硬编码用户ID，实际应该从JWT或Session中获取
-    private Long getCurrentUserId() {
-        return 1L; // TODO: 从认证信息中获取
+    private int getCurrentUserId() {
+        return 1; // TODO: 从认证信息中获取
     }
 
     /**
@@ -39,7 +39,7 @@ public class AnalysisController {
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size
     ) {
-        Long userId = getCurrentUserId();
+        int userId = getCurrentUserId();
         // 调用Service
         var result = analysisService.getLearningLogs(userId, targetType, actionType, startTime, endTime, page, size);
         return Result.success("获取成功", result);
@@ -50,7 +50,7 @@ public class AnalysisController {
      */
     @GetMapping("/all-logs")
     public Result<?> getAllLearningLogs(
-            @RequestParam(required = false) Long userId,
+            @RequestParam(required = false) int userId,
             @RequestParam(required = false) String targetType,
             @RequestParam(required = false) String actionType,
             @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime startTime,
@@ -67,7 +67,7 @@ public class AnalysisController {
      */
     @GetMapping("/all-logs/count")
     public Result<Long> countAllLearningLogs(
-            @RequestParam(required = false) Long userId,
+            @RequestParam(required = false) int userId,
             @RequestParam(required = false) String targetType,
             @RequestParam(required = false) String actionType,
             @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime startTime,
@@ -86,7 +86,7 @@ public class AnalysisController {
             @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime startTime,
             @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime endTime
     ) {
-        Long userId = getCurrentUserId();
+        int userId = getCurrentUserId();
         LearningStatsDTO stats = analysisService.getLearningStats(userId, startTime, endTime);
         return Result.success("获取成功", stats);
     }
@@ -96,7 +96,7 @@ public class AnalysisController {
      */
     @GetMapping("/user-stats/{userId}")
     public Result<LearningStatsDTO> getUserLearningStats(
-            @PathVariable Long userId,
+            @PathVariable int userId,
             @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime startTime,
             @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime endTime
     ) {
