+++ src/main/java/com/dd/ai_smart_course/dto/ConceptMasteryDTO.java
@@ -6,9 +6,9 @@ import java.time.LocalDateTime;
 
 @Data
 public class ConceptMasteryDTO {
-    private Long userId;
+    private int userId;
     private String username; // 添加用户名
-    private Long conceptId;
+    private int conceptId;
     private String conceptName; // 添加概念名称
     private Integer masteryLevel;
     private LocalDateTime lastUpdated;
