+++ src/main/java/com/dd/ai_smart_course/dto/LearningLogDTO.java
@@ -6,11 +6,11 @@ import java.time.LocalDateTime;
 
 @Data
 public class LearningLogDTO {
-    private Long id;
-    private Long userId;
+    private int id;
+    private int userId;
     private String username; // 添加用户名，方便前端展示
     private String targetType;
-    private Long targetId;
+    private int targetId;
     private String actionType;
     private LocalDateTime actionTime;
     private Integer duration;
